Move selectors to helpers directory
