+++ src/hooks/useWalletSectionsData.js
@@ -1,7 +1,5 @@
 import { useMemo } from 'react';
 import { useSelector } from 'react-redux';
-import { buildWalletSectionsSelector } from '../helpers/buildWalletSections';
-import { readableUniswapSelector } from '../hoc/uniswapLiquidityTokenInfoSelector';
 import useAccountAssets from './useAccountAssets';
 import useAccountSettings from './useAccountSettings';
 import useCoinListEditOptions from './useCoinListEditOptions';
@@ -9,6 +7,8 @@ import useIsWalletEthZero from './useIsWalletEthZero';
 import useSavingsAccount from './useSavingsAccount';
 import useSendableUniqueTokens from './useSendableUniqueTokens';
 import useShowcaseTokens from './useShowcaseTokens';
+import { buildWalletSectionsSelector } from '@rainbow-me/helpers/buildWalletSections';
+import { readableUniswapSelector } from '@rainbow-me/helpers/uniswapLiquidityTokenInfoSelector';
 
 export default function useWalletSectionsData() {
   const accountData = useAccountAssets();
